fix: correct metadata field_id type to u16, register microlensing example
- Change FIELD_* constants from u32 to u16 to match MetadataEntry.field_id type
- Add microlensing_detection example to Cargo.toml

+++ examples/rvf/examples/microlensing_detection.rs
@@ -28,10 +28,10 @@ use tempfile::TempDir;
 // Metadata field IDs
 // ---------------------------------------------------------------------------
 
-const FIELD_SURVEY: u32 = 0;
-const FIELD_EVENT_ID: u32 = 1;
-const FIELD_EINSTEIN_TIME: u32 = 2;
-const FIELD_ANOMALY_TYPE: u32 = 3;
+const FIELD_SURVEY: u16 = 0;
+const FIELD_EVENT_ID: u16 = 1;
+const FIELD_EINSTEIN_TIME: u16 = 2;
+const FIELD_ANOMALY_TYPE: u16 = 3;
 
 // ---------------------------------------------------------------------------
 // LCG helpers
